update types to use original exports
