+++ declarations/types/utils/toast_helper.d.ts
@@ -1,0 +1,9 @@
+import { TextStyle, ViewStyle } from 'react-native';
+export type ToastConfig = {
+    text: string;
+    icon?: string;
+    time?: number;
+    style?: ViewStyle;
+    slideTop?: boolean;
+    textStyle?: TextStyle;
+};
